feat(mako): deprecate contrib module on v2
