+++ docs/reference/contrib/mako.rst
@@ -1,5 +1,8 @@
 mako
 ====
 
+.. deprecated:: 2.22.0
+    Import from :mod:`litestar.plugins.mako` instead. This module will be removed in 3.0.0.
+
 .. automodule:: litestar.contrib.mako
     :members:
