+++ docs/reference/plugins/mako.rst
@@ -1,0 +1,6 @@
+====
+mako
+====
+
+.. automodule:: litestar.plugins.mako
+    :members:
